feat: Extension of XrmMockup
Created interface to enable extension of XrmMockup Core. The interface has one method, TriggerExtension, which for the moment will be executed after the Create,Update or Delete RequestMessage.
Extensions to the Core is set through XrmMockupSettings class.
Changed the build pipeline, to also build the Interface package. This enables extension project to only depend on the unfied interface, instead of whole XrmMockup.

For the moment, this feature is only available for XrmMockup365.